requirements.txt
anyio==3.6.1
asgiref==3.5.1
awscli==1.24.3
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.23.3
botocore==1.26.3
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.4
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.4
Django==4.0.4
docutils==0.15.2
entrypoints==0.3
fastapi==0.78.0
filelock==3.6.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.28.2
h11==0.13.0
idna==3.3
ipykernel==6.5.1
ipython==7.29.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mibian==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.2
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
plotly==5.8.0
plotly-express==0.4.1
prompt-toolkit==3.0.22
psycopg2==2.9.3
pyasn1==0.4.8
pydantic==1.9.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
#pypiwin32==219
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
scipy==1.8.1
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.19.1
statsmodels==0.13.2
tenacity==8.0.1
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.1.2
zope.interface==5.4.0
